[TASK] Drop SchedulerTaskLinkBuilder interface
