+++ Tests/Functional/Provider/Scheduler/SchedulerProviderFunctionalTestCase.php
@@ -19,13 +19,16 @@ namespace mteu\Monitoring\Tests\Functional\Provider\Scheduler;
 
 use mteu\Monitoring\Configuration\Provider\SchedulerProviderConfiguration;
 use mteu\Monitoring\Provider\Scheduler\SchedulerProvider;
+use mteu\Monitoring\Provider\Scheduler\SchedulerTaskLinkBuilder;
 use mteu\Monitoring\Tests\Functional\MonitoringFunctionalTestCase;
-use mteu\Monitoring\Tests\Unit\Fixtures\Clock\FrozenClock;
 use mteu\Monitoring\Tests\Unit\Fixtures\Scheduler\InMemorySchedulerHeartbeat;
 use mteu\Monitoring\Tests\Unit\Fixtures\Scheduler\InMemorySchedulerTaskRepository;
-use mteu\Monitoring\Tests\Unit\Fixtures\Scheduler\InMemoryTaskLinkBuilder;
 use Psr\Log\NullLogger;
 use Symfony\Component\Clock\MockClock;
+use TYPO3\CMS\Backend\Routing\Exception\RouteNotFoundException;
+use TYPO3\CMS\Backend\Routing\Router;
+use TYPO3\CMS\Backend\Routing\UriBuilder;
+use TYPO3\CMS\Core\Http\Uri;
 
 /**
  * ProviderFunctionalTestCase.
@@ -49,8 +52,28 @@ abstract class SchedulerProviderFunctionalTestCase extends MonitoringFunctionalT
             new InMemorySchedulerTaskRepository(),
             new InMemorySchedulerHeartbeat(self::NOW - 60),
             new MockClock((new \DateTimeImmutable())->setTimestamp(self::NOW)),
-            new InMemoryTaskLinkBuilder(null),
+            $this->createLinkBuilder(null),
             new NullLogger(),
         );
+    }
+
+    /**
+     * Builds the real link builder with a stubbed routing stack: when no base
+     * URI is given the UriBuilder fails to resolve a route, so no link is built.
+     */
+    private function createLinkBuilder(?string $baseUri): SchedulerTaskLinkBuilder
+    {
+        $router = self::createStub(Router::class);
+        $router->method('hasRoute')->willReturn(false);
+
+        $uriBuilder = self::createStub(UriBuilder::class);
+
+        if ($baseUri === null) {
+            $uriBuilder->method('buildUriFromRoute')->willThrowException(new RouteNotFoundException());
+        } else {
+            $uriBuilder->method('buildUriFromRoute')->willReturn(new Uri($baseUri));
+        }
+
+        return new SchedulerTaskLinkBuilder($router, $uriBuilder, new NullLogger());
     }
 }
